+++ database/migrations/2024_10_08_081009_create_guide_languages_table.php
@@ -13,13 +13,12 @@ return new class extends Migration
     {
         Schema::create('guide_languages', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('guide_id');
-            $table->unsignedInteger('language_id');
+            $table->unsignedBigInteger('guide_id');
+            $table->unsignedBigInteger('language_id');
             $table->timestamps();
 
             $table->foreign('guide_id')->references('id')->on('guides')->onDelete('cascade');
             $table->foreign('language_id')->references('id')->on('languages')->onDelete('cascade');
-      
         });
     }
 
@@ -29,14 +28,11 @@ return new class extends Migration
     public function down(): void
     {
        Schema::table('guide_languages', function (Blueprint $table) {
-            
             $table->dropForeign(['guide_id']);
             $table->dropForeign(['language_id']);
             
             $table->dropColumn('guide_id');
             $table->dropColumn('language_id');
-
-
         });
     }
 };
